fix: add CSP support for image data-blobs
It is very expensive in terms of time and amount of data
to load very small images via HTTP. Thus browsers do
support data-URLs for images. Utilizing this, small
images can be transmitted along the JavaScript, CSS or HTML.

The previous setting for the Content-Security-Policy header
did not permit such data-URLs, as these must be explicitly
named to work with major browsers like Chrome.
This commit adds necessary setting to support data-URLs in
these browsers.

see:
* https://stackoverflow.com/questions/18447970/content-security-policy-data-not-working-for-base64-images-in-chrome-28#18449556
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/Content-Security-Policy/img-src
* https://developer.mozilla.org/en-US/docs/Web/HTTP/Basics_of_HTTP/Data_URIs