feature:(tenant) support tenant 0 white list of URLs
For tenant with ID, only as single URL was allowed, previously.
It was configured as "Identity:Tenants:TenantSelectorFallbackUrl".

Now, the configuration may contain a list of allowed URLs for this
tenant, which can be configured with configuration setting
"Identity:Tenants:TenantSelectorWhitelistUrls". The JSON is an array
of string values.

All other URLs are redirected to the fallback URL.
The values do not support regular expressions - yet.