fix:(CSP) allow *.smint.io for iframe parents
Office365 requires the use of an iFrame layer wrapper from
the same domain. So the web page must be able to be loaded
in an iframe from the same (*.smint.io) main domain.

All this to overcome restrictions of Aadd-in manifest settings,
that do not allow the use of wildcards with allowed domains.

!!Beware!! that the local domain for the development machines
is not included in this new settings. So each developer must
add it on its own.