feature:(launch) enable additional configuration with Launcher
Additional configuration can be used to customize the configuration
at compile time. There are some configurations that do not change
at runtime but are still necessary for the application.

The additional configuration is applied to the startup config and
the application config. The former is dropped once the application
has been started.