refact:(filter) add configurable HTTP headers with SecurityHeadersAttribute
Previously all security headers, like Content-Security-Policy, were hard
coded in the code. For a library this is very bad and a no-brainer as
it ties all usages of the library to a specific set-up.

This has been fixed now. The headers to add can be configured now. The
global `IConfiguration` instance is used to hold header configuration.

Nevertheless, previous behaviour is maintained, if no configuration has
been set.