refact:(filter) make HTTP header customizable with tag names
A tag name can be set to the HTTP filter for each class this filter
is applied to. So classes may use variants of the HTTP headers in
case they need it.

Configuration values are read from a set of configuration that contain
the provided tag name. These values override the default values,
which are stored unsing the tag name value `default`.

To use a custom tag name, follow this pattern:
```
    [TypeFilter(typeof(SecurityHeadersAttribute), Arguments = new object[] {"ui-as-plugin"})]
    public class PageWithSecurityHeaders : PageModel
    {
    }
```