fix:(filter) apply HTTP headers to redirection result too
Because the security headers are missing from redirects,
authentication fails. The authentication checks the user
credentials and the redirects to the root page. Along
the redirection headers, proper cookie headers are set
to authorize the user to access the root page.

If the security headers are missing, then the cookies
are not accepted if the web page is running as plugin
inside Office365-for-the-web iFrame. Then the authentication
is successfully performed but the user autorization is immediately
"forgotten". Hence the security headers are needed for
redirects, too.