feature:(web) extract JS/CSS from static html file of requested page template
Although all HTTP pages are served via Razor with .NET MVC, sometimes
it is supportive to load a template HTML file, extract script and CSS
includes and add these to the result of the Razor page.

This system enables the Razor pages to wrap static HTML files, using their
includes but overwriting the HTML parts.

If no such HTML file is found, then the previous SPA manifest
based solution is used as fallback.