refact:(web) provide flag "UseSpa" to public
Provding the flag `UseSpa` to the public enables
decendant classes to perform special initialisation
for SPA apps without the need to know the detailed
internals of the flag. It will avoid using another
custom flag for the same purpose.

SPA apps seems different enough and important enough
to provide this flag to the public.