fix:(web) use enpoint path as HTML template fallback
This is vital for SPA pages!!

Using wildcards in routes, the URL path might be a
"virtual" path instead of a real file path. In such cases
the HTML template file can not be found, although the
endpoint has one. So, the endpoint path is used as
a fallback to find a HTML template file to use.

This is necessary and vital if there is some sort of
"catch-all" page, like "/Index" endpoint to serve
a Single-Page-Application (SPA) page.