refact:(download-proxy) do not convert to/from JSON of parameters
Previously all parameters for the download proxy have been protected
by .NET core "data protection" facility, which protects string values.
On the other hand there is fear that this facility will increase the
value by an undetermined amount of bytes and thus might exceed
maximum query limit. Hence the parameters are not protected but used
directly on the URL query. Only the hash value of these parameters
is data protected now.

Because of the initial protection of all parameters, these parameters
have been stored in an object and converted to and from JSON. The JSON
string has been data protected.

Now, that no object is used anymore, no JSON is necessary anymore.
So this is removed.